#60 Roles de Subject

+++ EsfotFeedback/app/Http/Controllers/SubjectController.php
@@ -11,11 +11,13 @@ class SubjectController extends Controller
 {
     public function index()
     {
+        $this->authorize('viewAny', Subject::class);
         return new SubjectCollection(Subject::paginate(25));
     }
 
     public function show(Subject $subject)
     {
+        $this->authorize('view', $subject);
         return response()->json(new SubjectResource($subject),200);
     }
 
@@ -27,12 +29,14 @@ class SubjectController extends Controller
 
     public function update(Request $request, Subject $subject)
     {
+        $this->authorize('update', $subject);
         $subject->update($request->all());
         return response()->json($subject, 200);
     }
 
     public function delete(Subject $subject)
     {
+        $this->authorize('delete', $subject);
         $subject->delete();
         return response()->json(null, 204);
     }
